refactored cast list so simple slider is a reusable component
